[IMBH-11] Jobs de descarga y filtrado
Se ha eliminado la carpeta de executables y se ha guardado los jobs en la carpeta jobs

+++ src/hyper_velocity_stars_detection/jobs/google_jobs/filter_data.py
@@ -5,10 +5,10 @@ from tempfile import TemporaryDirectory
 
 from astroquery.simbad import Simbad
 from google.cloud import storage
-from project_vars import get_params
 from tqdm import tqdm
-from utils import upload_folder_to_gcs
 
+from hyper_velocity_stars_detection.jobs.google_jobs.utils import upload_folder_to_gcs
+from hyper_velocity_stars_detection.jobs.utils import get_params
 from hyper_velocity_stars_detection.sources.source import AstroObject, AstroObjectData
 
 Simbad.SIMBAD_URL = "http://simbad.u-strasbg.fr/simbad/sim-id"
@@ -38,21 +38,23 @@ if __name__ == "__main__":
         with TemporaryDirectory() as temp_path:
             local_path = os.path.join(temp_path, file_name)
             logging.info(f"Descargando {file_name} desde {folder}...")
-            blob.download_to_filename(local_path)
             name = file_name.split("_")[1]
             radio_scale = int(file_name.split("_")[2].replace("r", "").replace(".vot", ""))
-            astro_object = AstroObject.get_object(name)
-            _ = astro_object.read_object(temp_path, file_name)
-            pm_kms = None
-            if radio_scale > 1:
-                pm_kms = args.pm_kms
-            astro_data = AstroObjectData.load_data_from_object(
-                astro_object, radio_scale, pmra_kms_min=pm_kms, pmdec_kms_min=pm_kms
-            )
-            path_temp = os.path.join(temp_path, name)
-            astro_data.save(path_temp)
-            upload_folder_to_gcs(project_id, bucket_name, path_temp, name)
-            os.remove(local_path)  # Borrar archivo local
+            name_blobs = list(bucket.list_blobs(prefix=name))
+            if len(name_blobs) < 2:
+                blob.download_to_filename(local_path)
+                astro_object = AstroObject.get_object(name)
+                _ = astro_object.read_object(temp_path, file_name)
+                pm_kms = None
+                if radio_scale > 1:
+                    pm_kms = args.pm_kms
+                astro_data = AstroObjectData.load_data_from_object(
+                    astro_object, radio_scale, pmra_kms_min=pm_kms, pmdec_kms_min=pm_kms
+                )
+                path_temp = os.path.join(temp_path, name)
+                astro_data.save(path_temp)
+                upload_folder_to_gcs(project_id, bucket_name, path_temp, name)
+                os.remove(local_path)
 
         logging.info(f"Borrando {file_name} del bucket...")
         blob.delete()
